refactor: move document array io to mixin
